common: Remove the driver abstraction from the wall clock
All sunxi SoCs that have the AR100 CPU also contain a 64-bit counter
(sometimes called "CNT64" in the documentation) in the R_CPUCFG MMIO
area. Instead of providing access to this device through a driver, put
the function for reading it in common code. This allows it to be used
at all times (including during driver model initialization) for delays
and timeouts, and logging.

In case future SoCs do not contain the same CNT64 hardware at the same
offsets, declare the counter_read function as weak so it can be
overridden in platform code (i.e. platform/newplatform/counter.c).

Remove the old driver class, all existing drivers, and all references to
the drivers in platform code.

Signed-off-by: Samuel Holland <samuel@sholland.org>